+++ ft_printf_c.c
@@ -1,6 +1,6 @@
 #include "ft_printf.h"
 
-void    ft_c_right(t_flags *prt, va_list args, int *count_char)
+void    ft_c_right(t_flags *prt, va_list args, size_t *count_char)
 {
     int n;
 
@@ -13,7 +13,7 @@ void    ft_c_right(t_flags *prt, va_list args, int *count_char)
     }
 }
 
-void    ft_c_left(t_flags *prt, char q, va_list args, int *count_char)
+void    ft_c_left(t_flags *prt, char q, va_list args, size_t *count_char)
 {
     int n;
